update
Old calculators moved to 'deprecated' folder.
Added better styled calculators with time tracking. Now reqursive actions are really reqursive.

+++ fork_join_framework/src/main/java/SumCalculator.java
@@ -1,4 +1,4 @@
 public interface SumCalculator {
     void compute();
-    Long getResult();
+    Integer getResult();
 }
